feat: Naive solution
Implemented a naive with maximum O(2N + 2M) complexity where N
is the size of list1 and M is the size of list2.
The maximum time complexity for this is O(4N) where N is the size
of the largest list
The space complexity for this is O(N + M).
This is the worst case and most naive solution possible.